vivarium/requirements.txt
# ============================================================
# Vivarium Monitoring Pipeline — Requirements
# Python 3.11+
# ============================================================

# --- Core API framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9          # multipart/form-data uploads

# --- Configuration ---
PyYAML>=6.0.1
dotmap>=1.3.30                   # attribute-access dict (config.key syntax)
python-dotenv>=1.0.1             # load .env for local dev

# --- Database ---
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0                  # async PostgreSQL driver

# --- Computer vision (core) ---
opencv-python>=4.9.0

# --- Detector — YOLOv8-World ---
ultralytics>=8.2.0

# --- Detector — Grounding DINO (optional, comment out if not needed) ---
transformers>=4.40.0
torch>=2.2.0

# --- Measurer — CLIP ---
openai-clip>=1.0.1               # pip install openai-clip (not 'clip')
# Note: torch already listed above

# --- Measurer — custom ONNX classifier (optional) ---
onnxruntime>=1.18.0

# --- Notifiers ---
aiohttp>=3.9.0                   # Telegram + webhook async HTTP

# --- Logging ---
loguru>=0.7.2

# --- Image processing ---
Pillow>=10.3.0

# --- Cloud image stores (optional, install only what you need) ---
# aioboto3>=12.3.0               # AWS S3
# google-cloud-storage>=2.16.0   # Google Cloud Storage

# --- Testing ---
pytest>=8.1.0
pytest-asyncio>=0.23.6
pytest-mock>=3.14.0
httpx>=0.27.0                    # async test client for FastAPI

# --- Rate limiting (optional, plug into FastAPI middleware) ---
# slowapi>=0.1.9
